+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Changes:
 
 - Fix selected colour in colour changing dialog
+- Support for deleting multiple cards at once
 
 ## v2.0.4 (2021-07-27)